Paylasima_Hazir/Antigravity 2/Projeler/Satis_Notifier/requirements.txt
google-api-python-client==2.114.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
requests>=2.28.0
pytz>=2023.3
